Adjust file contents a little.
