+++ servant-foreign/src/Servant/Foreign/Internal.hs
@@ -16,35 +16,7 @@
 
 -- | Generalizes all the data needed to make code generation work with
 -- arbitrary programming languages.
-module Servant.Foreign
-  ( HasForeign(..)
-  , Segment(..)
-  , SegmentType(..)
-  , FunctionName
-  , QueryArg(..)
-  , HeaderArg(..)
-  , ArgType(..)
-  , Req
-  , captureArg
-  , defReq
-  , concatCase
-  , snakeCase
-  , camelCase
-  -- lenses
-  , argType
-  , argName
-  , isCapture
-  , funcName
-  , path
-  , reqUrl
-  , reqBody
-  , reqHeaders
-  , reqMethod
-  , segment
-  , queryStr
-  -- re-exports
-  , module Servant.API
-  ) where
+module Servant.Foreign.Internal where
 
 import           Control.Lens (makeLenses, (%~), (&), (.~), (<>~))
 import qualified Data.Char    as C
